+++ backend/internal/query/sql/common/query_stats_enabled/query_stats_enabled.tmpl.sql
@@ -1,1 +1,1 @@
-SELECT COUNT(*) > 0 AS count FROM pg_extension WHERE extname = 'pg_stat_statements'
+SELECT COUNT(*) > 0 AS count FROM pg_extension WHERE extname = '{{ .Extension }}'